Adding authentication support with redis/eth gateways

+++ api/usecases/delete_comment.go
@@ -2,6 +2,7 @@ package usecases
 
 import (
 	"context"
+	"errors"
 
 	"github.com/daochanio/backend/api/gateways"
 	"github.com/daochanio/backend/common"
@@ -22,10 +23,19 @@ type DeleteCommentInput struct {
 	DeleterAddress string `validate:"eth_addr"`
 }
 
-// TODO: check if the comment belongs to the user (or is a moderate once supported) before deleting
 func (u *DeleteCommentUseCase) Execute(ctx context.Context, input DeleteCommentInput) error {
 	if err := common.ValidateStruct(input); err != nil {
 		return err
+	}
+
+	comment, err := u.dbGateway.GetCommentById(ctx, input.Id)
+
+	if err != nil {
+		return err
+	}
+
+	if comment.Address() != input.DeleterAddress {
+		return errors.New("comment does not belong to the user")
 	}
 
 	return u.dbGateway.DeleteComment(ctx, input.Id)
